+++ plugins/catalog-backend/src/ingestion/HigherOrderOperations.ts
@@ -132,7 +132,7 @@ export class HigherOrderOperations implements HigherOrderOperation {
         await this.locationsCatalog.logUpdateSuccess(location.id, undefined);
       } catch (e) {
         this.logger.warn(
-          `Failed to refresh location ${location.type}:${location.target}, ${e}`,
+          `Failed to refresh location ${location.type}:${location.target}, ${e.stack}`,
         );
         await this.locationsCatalog.logUpdateFailure(location.id, e);
       }
@@ -186,7 +186,7 @@ export class HigherOrderOperations implements HigherOrderOperation {
 
     for (const item of readerOutput.errors) {
       this.logger.warn(
-        `Failed item in location ${item.location.type}:${item.location.target}, ${item.error}`,
+        `Failed item in location ${item.location.type}:${item.location.target}, ${item.error.stack}`,
       );
     }