+++ plugins/catalog-backend/src/service/CatalogBuilder.ts
@@ -37,23 +37,23 @@ import {
 import { DatabaseManager } from '../database';
 import {
   AnnotateLocationEntityProcessor,
+  BuiltinKindsEntityProcessor,
   CatalogProcessor,
   CodeOwnersProcessor,
   FileReaderProcessor,
   GithubOrgReaderProcessor,
   HigherOrderOperation,
   HigherOrderOperations,
+  LdapOrgReaderProcessor,
   LocationReaders,
   LocationRefProcessor,
-  OwnerRelationProcessor,
+  MicrosoftGraphOrgReaderProcessor,
   PlaceholderProcessor,
   PlaceholderResolver,
   StaticLocationProcessor,
   UrlReaderProcessor,
 } from '../ingestion';
 import { CatalogRulesEnforcer } from '../ingestion/CatalogRules';
-import { BuiltinKindsEntityProcessor } from '../ingestion/processors/BuiltinKindsEntityProcessor';
-import { LdapOrgReaderProcessor } from '../ingestion/processors/LdapOrgReaderProcessor';
 import {
   jsonPlaceholderResolver,
   textPlaceholderResolver,
@@ -278,10 +278,10 @@ export class CatalogBuilder {
         new FileReaderProcessor(),
         GithubOrgReaderProcessor.fromConfig(config, { logger }),
         LdapOrgReaderProcessor.fromConfig(config, { logger }),
+        MicrosoftGraphOrgReaderProcessor.fromConfig(config, { logger }),
         new UrlReaderProcessor({ reader, logger }),
         new CodeOwnersProcessor({ reader }),
         new LocationRefProcessor(),
-        new OwnerRelationProcessor(),
         new AnnotateLocationEntityProcessor(),
       );
     }