+++ plugins/scaffolder/src/types.ts
@@ -13,6 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+export type JobStatus = 'PENDING' | 'STARTED' | 'COMPLETED' | 'FAILED';
 export type Job = {
   id: string;
   metadata: {
@@ -20,8 +22,9 @@ export type Job = {
     values: any;
     remoteUrl?: string;
     token?: string;
+    catalogInfoUrl?: string;
   };
-  status: 'PENDING' | 'STARTED' | 'COMPLETED' | 'FAILED';
+  status: JobStatus;
   stages: Stage[];
   error?: Error;
   token: string;
@@ -30,7 +33,7 @@ export type Job = {
 export type Stage = {
   name: string;
   log: string[];
-  status: 'PENDING' | 'STARTED' | 'COMPLETED' | 'FAILED';
+  status: JobStatus;
   startedAt: string;
   endedAt?: string;
 };