+++ plugins/jenkins/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @backstage/plugin-jenkins
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [475fc0aaa]
+- Updated dependencies [1166fcc36]
+- Updated dependencies [1185919f3]
+  - @backstage/core@0.3.2
+  - @backstage/catalog-model@0.3.0
+  - @backstage/plugin-catalog@0.2.3
+
+## 0.3.0
+
+### Minor Changes
+
+- a41730c6e: Add tooltip for Jenkins rerun button
+
+### Patch Changes
+
+- Updated dependencies [7b37d65fd]
+- Updated dependencies [4aca74e08]
+- Updated dependencies [e8f69ba93]
+- Updated dependencies [0c0798f08]
+- Updated dependencies [0c0798f08]
+- Updated dependencies [199237d2f]
+- Updated dependencies [6627b626f]
+- Updated dependencies [4577e377b]
+- Updated dependencies [2d0bd1be7]
+  - @backstage/core@0.3.0
+  - @backstage/theme@0.2.1
+  - @backstage/plugin-catalog@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes